UI test cleanup: Extract for_kv_map lint tests
