+++ tests/ui/for_kv_map.stderr
@@ -1,0 +1,54 @@
+error: you seem to want to iterate on a map's values
+  --> $DIR/for_kv_map.rs:18:19
+   |
+LL |     for (_, v) in &m {
+   |                   ^^
+   |
+   = note: `-D clippy::for-kv-map` implied by `-D warnings`
+help: use the corresponding method
+   |
+LL |     for v in m.values() {
+   |         ^    ^^^^^^^^^^
+
+error: you seem to want to iterate on a map's values
+  --> $DIR/for_kv_map.rs:23:19
+   |
+LL |     for (_, v) in &*m {
+   |                   ^^^
+help: use the corresponding method
+   |
+LL |     for v in (*m).values() {
+   |         ^    ^^^^^^^^^^^^^
+
+error: you seem to want to iterate on a map's values
+  --> $DIR/for_kv_map.rs:31:19
+   |
+LL |     for (_, v) in &mut m {
+   |                   ^^^^^^
+help: use the corresponding method
+   |
+LL |     for v in m.values_mut() {
+   |         ^    ^^^^^^^^^^^^^^
+
+error: you seem to want to iterate on a map's values
+  --> $DIR/for_kv_map.rs:36:19
+   |
+LL |     for (_, v) in &mut *m {
+   |                   ^^^^^^^
+help: use the corresponding method
+   |
+LL |     for v in (*m).values_mut() {
+   |         ^    ^^^^^^^^^^^^^^^^^
+
+error: you seem to want to iterate on a map's keys
+  --> $DIR/for_kv_map.rs:42:24
+   |
+LL |     for (k, _value) in rm {
+   |                        ^^
+help: use the corresponding method
+   |
+LL |     for k in rm.keys() {
+   |         ^    ^^^^^^^^^
+
+error: aborting due to 5 previous errors
+
